requirements.txt
# Core
numpy==1.24.3
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74
pillow==10.0.0

# Deep Learning
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
ultralytics==8.0.200
mediapipe==0.10.5

# Tracking
scikit-learn==1.3.0
scipy==1.11.2

# Web Framework
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0

# Database
# sqlite3 is built-in

# IoT (Raspberry Pi only)
# Uncomment when deploying to Raspberry Pi:
# RPi.GPIO==0.7.1
twilio==8.8.0
firebase-admin==6.2.0

# Utilities
pandas==2.0.3
matplotlib==3.7.2
tqdm==4.66.1
pytest==7.4.0
pytest-cov==4.1.0
